refactor: OrderLineItem JPA 적용
