+++ src/main/java/kitchenpos/domain/OrderLineItem.java
@@ -1,9 +1,25 @@
 package kitchenpos.domain;
 
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+
+@Entity
 public class OrderLineItem {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long seq;
+
+    @Column(nullable = false)
     private Long orderId;
+
+    @Column(nullable = false)
     private Long menuId;
+
+    @Column(nullable = false)
     private long quantity;
 
     public Long getSeq() {
