+++ src/main/java/kitchenpos/repository/OrderLineItemRepository.java
@@ -1,0 +1,10 @@
+package kitchenpos.repository;
+
+import java.util.List;
+import kitchenpos.domain.OrderLineItem;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderLineItemRepository extends JpaRepository<OrderLineItem, Long> {
+
+    List<OrderLineItem> findAllByOrderId(Long orderId);
+}
